pyproject.toml
[project]
name = "timeline-generator-mcp"
version = "0.1.0"
description = "A CLI tool and MCP server to generate beautiful timeline images and animated GIFs from milestone data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Timeline Generator Team"}
]
keywords = ["timeline", "visualization", "gif", "animation", "cli", "mcp", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "typer>=0.9.0",
    "pycairo>=1.25.0",
    "pillow>=10.0.0",
    "moviepy>=1.0.3",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "mcp>=1.0.0",
    "fastmcp>=0.4.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
timeline-gen = "timeline_generator.cli:app"
timeline-mcp = "timeline_generator.fastmcp_server:main"
timeline-mcp-http = "timeline_generator.fastmcp_server:main_http"
timeline-mcp-legacy = "timeline_generator.mcp_server:main"

[project.urls]
Homepage = "https://github.com/kbichave/timeline-generator-mcp"
Repository = "https://github.com/kbichave/timeline-generator-mcp"
Issues = "https://github.com/kbichave/timeline-generator-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["timeline_generator"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

